Horture: refactor architecture local client <-> server comm + GUI

+++ src/Horture/Events.hs
@@ -15,7 +15,7 @@ import Horture.Logging
 import Horture.Scene
 import Horture.State
 
-pollHortureEvents :: (HortureLogger (Horture l hdl)) => Double -> Double -> Scene -> Horture l hdl (Maybe Scene)
+pollHortureEvents :: (HortureLogger (Horture l hdl)) => Float -> Float -> Scene -> Horture l hdl (Maybe Scene)
 pollHortureEvents timeNow dt s = do
   asks _eventChan
     >>= liftIO . tryReadChan
@@ -26,12 +26,12 @@ pollHortureEvents timeNow dt s = do
       _otherwise -> do
         return (Just s)
 
-handleHortureEvent :: Double -> Double -> Event -> Scene -> Horture l hdl (Maybe Scene)
+handleHortureEvent :: Float -> Float -> Event -> Scene -> Horture l hdl (Maybe Scene)
 handleHortureEvent timeNow dt (EventEffect n eff) s = do
   gets (^. eventList) >>= liftIO . Ringbuffer.append (PastEvent timeNow n eff)
   Just <$> applyEffect timeNow dt s eff
 handleHortureEvent _ _ (EventCommand Exit) _ = return Nothing
 handleHortureEvent _ _ (EventCommand _cmd) _ = throwError $ HE "unimplemented"
 
-applyEffect :: Double -> Double -> Scene -> Effect -> Horture l hdl Scene
+applyEffect :: Float -> Float -> Scene -> Effect -> Horture l hdl Scene
 applyEffect timeNow dt s eff = return $ apply timeNow dt eff s
